Minor updates for type consistency. Added stricter catch expressions. Fixed data corruption due to mismatch in writeSectionToFile and readSectionsFromFile.

+++ BACC-Portal/src/main/java/model/academics/CourseSection.java
@@ -42,11 +42,11 @@ public class CourseSection {
         return price;
     }
 
-    public double getTotalCapacity() {
+    public int getTotalCapacity() {
         return totalCapacity;
     }
 
-    public double getCurrentCapacity() {
+    public int getCurrentCapacity() {
         return currentCapacity;
     }
